requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Async database and Redis
asyncpg==0.29.0
aioredis==2.0.1
aiohttp==3.9.0

# Task scheduling
APScheduler==3.10.4

# Configuration and data handling
PyYAML==6.0.1
structlog==23.2.0

# Email functionality
aiosmtplib==3.0.0

# Monitoring and metrics
prometheus-client==0.19.0

# Additional utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.2

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
# asyncio
# asyncpg==0.28.0
# aioredis==2.0.1
# aiohttp==3.8.5
# apscheduler==3.10.4
# structlog==23.1.0
# prometheus-client==0.17.1
# psycopg2-binary==2.9.7
# redis==4.6.0
# python-dotenv==1.0.0
# pydantic==2.1.1
# fastapi==0.101.1
# uvicorn==0.23.2